requirements.txt
numpy==1.18.5
Keras==2.4.3
streamlit==0.69.2
Pillow==8.0.0
tensorflow==2.3.1
